Small Fixes to Get-FileEncoding
This change allows the Get-FileEncoding cmdlet to run on PowerShell core
and thus on Linux.

The encoding parameter to get content has changed in core so that 'byte'
is no longer valid.

Unfortunately I had to resort to a stream reader since Get-Content likes
to strip a BOM if it exists. The -raw switch only ignores line endings.

The alias for Sort-Object to simply 'Sort' has also been removed in
the latest versions of PS Core. This change also removes the 'Select'
alias, even though it isn't technically removed yet, because there is
some discussion around removing all of these aliases in the default
PS Core environments, so best to get rid of it now.

https://github.com/PowerShell/PowerShell/issues/5870